requirements.txt
sigmatools>=0.23.0
ruamel.yaml>=0.18.0
pyyaml>=6.0
requests>=2.31.0
urllib3>=2.0.0
jsonschema>=4.21.0
jinja2>=3.1.3
pyspellchecker>=0.8.1
packaging>=24.0
pytest>=8.0.0
sigma-cli
elasticsearch>=8.0.0
pySigma-backend-elasticsearch
